backend/requirements.txt
fastapi==0.103.0
uvicorn[standard]==0.23.2
pydantic==2.3.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
requests==2.31.0
pandas==2.0.3
numpy==1.24.3
plotly==5.15.0
yfinance==0.2.20
alpha-vantage==2.3.1
finnhub-python==2.4.18
websocket-client==1.6.1
aiofiles==23.2.1
aiomysql==0.2.0
alembic==1.11.1
sqlalchemy==2.0.19
redis==4.6.0
celery==5.3.1
python-dateutil==2.8.2
schedule==1.2.0
email-validator==2.0.0
httpx==0.24.1
websockets==11.0.3
typing-extensions==4.7.1
pycryptodome==3.18.0
bcrypt==4.0.1
cryptography==41.0.3
polygon-api-client==1.12.3
twelvedata==1.0.0
iexfinance==0.5.0
pycoingecko==3.1.0
requests-cache==1.1.1
aiohttp==3.8.5
ccxt==4.1.13
ta==0.10.2
scikit-learn==1.3.0
statsmodels==0.14.0 